Scaffold project cards and move stuff around
