+++ front/app/containers/Admin/inspirationHub/ProjectCards/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { Box } from '@citizenlab/cl2-component-library';
+
+const ProjectCards = () => {
+  return <Box>TODO</Box>;
+};
+
+export default ProjectCards;
