+++ front/app/containers/Admin/inspirationHub/ProjectDrawer/Phase.tsx
@@ -8,7 +8,7 @@ import { ProjectLibraryProjectData } from 'api/project_library_projects/types';
 import { useIntl } from 'utils/cl-intl';
 import { parseBackendDateString } from 'utils/dateUtils';
 
-import MethodLabel from '../MethodLabel';
+import MethodLabel from '../components/MethodLabel';
 import { useLocalizeProjectLibrary } from '../utils';
 
 import ExternalLink from './ExternalLink';
