+++ front/app/containers/Admin/inspirationHub/components/MethodLabel.tsx
